requirements.txt
fastapi
uvicorn
aiofiles
python-dotenv
aiogram
pony
pytz
apscheduler
yookassa
pycryptodome